https://telecominfraproject.atlassian.net/browse/WIFI-13147
