+++ src/AP_WS_Reactor_Pool.h
@@ -12,7 +12,6 @@
 #include <Poco/Data/SessionPool.h>
 #include "framework/utils.h"
 #include <StorageService.h>
-#include <AP_WS_Locked_Session.h>
 
 namespace OpenWifi {
 
@@ -33,16 +32,14 @@ namespace OpenWifi {
 			Threads_.reserve(NumberOfThreads_);
 			Logger_.information(fmt::format("WebSocket Processor: starting {} threads.", NumberOfThreads_));
 			for (uint64_t i = 0; i < NumberOfThreads_; ++i) {
-				auto NewReactor = std::make_unique<Poco::Net::SocketReactor>();
+				auto NewReactor = std::make_shared<Poco::Net::SocketReactor>();
 				auto NewThread = std::make_unique<Poco::Thread>();
 				NewThread->start(*NewReactor);
 				std::string ThreadName{"ap:react:" + std::to_string(i)};
 				Utils::SetThreadName(*NewThread, ThreadName.c_str());
 				Reactors_.emplace_back(std::move(NewReactor));
 				Threads_.emplace_back(std::move(NewThread));
-				LockedDbSession	S{ 	.Session = std::make_unique<Poco::Data::Session>(Poco::Data::Session(StorageService()->Pool().get())),
-								  	.Mutex = std::make_unique<std::mutex>()	};
-				DbSessions_.emplace_back(std::move(S));
+				DbSessions_.emplace_back(std::make_shared<LockedDbSession>());
 			}
 			Logger_.information(fmt::format("WebSocket Processor: {} threads started.", NumberOfThreads_));
 		}
@@ -58,20 +55,20 @@ namespace OpenWifi {
 			DbSessions_.clear();
 		}
 
-		std::pair<Poco::Net::SocketReactor *, LockedDbSession *> NextReactor() {
+		std::pair<std::shared_ptr<Poco::Net::SocketReactor>, std::shared_ptr<LockedDbSession> > NextReactor() {
 			std::lock_guard Lock(Mutex_);
 			NextReactor_++;
 			NextReactor_ %= NumberOfThreads_;
-			return std::make_pair(Reactors_[NextReactor_].get(), &DbSessions_[NextReactor_]);
+			return std::make_pair(Reactors_[NextReactor_], DbSessions_[NextReactor_]);
 		}
 
 	  private:
 		std::mutex Mutex_;
 		uint64_t NumberOfThreads_;
 		uint64_t NextReactor_ = 0;
-		std::vector<std::unique_ptr<Poco::Net::SocketReactor>> 	Reactors_;
+		std::vector<std::shared_ptr<Poco::Net::SocketReactor>> 	Reactors_;
 		std::vector<std::unique_ptr<Poco::Thread>> 				Threads_;
-		std::vector<LockedDbSession> 							DbSessions_;
+		std::vector<std::shared_ptr<LockedDbSession>>			DbSessions_;
 		Poco::Logger &Logger_;
 
 	};
